Warn if navigating away from unsaved changes on store payment settings
